requirements.txt
Django==3.2.7
djangorestframework==3.13.1
geopy==2.2.0
googlemaps==4.6.0
pandas==1.3.3
SQLAlchemy==1.4.35
